% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{acp_call_action}
\alias{acp_call_action}
\title{Call an ACP action endpoint}
\usage{
acp_call_action(client, action_name, body = list())
}
\arguments{
\item{client}{ACP client object}

\item{action_name}{action name without the \verb{/actions/} prefix}

\item{body}{request payload}
}
\value{
parsed ACP response
}
\description{
Call an ACP action endpoint
}
