% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{acp_call_flow}
\alias{acp_call_flow}
\title{Call an ACP flow endpoint}
\usage{
acp_call_flow(client, flow_name, body = list())
}
\arguments{
\item{client}{ACP client object}

\item{flow_name}{flow name without the \verb{/flows/} prefix}

\item{body}{request payload}
}
\value{
parsed ACP response
}
\description{
Call an ACP flow endpoint
}
