% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.R
\name{acp_cohort_methods_intent_split}
\alias{acp_cohort_methods_intent_split}
\title{Call cohort methods intent split flow}
\usage{
acp_cohort_methods_intent_split(client, study_intent)
}
\arguments{
\item{client}{ACP client object}

\item{study_intent}{non-empty study intent string}
}
\value{
parsed ACP response
}
\description{
Call cohort methods intent split flow
}
