% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_api.R
\name{acp_connect}
\alias{acp_connect}
\title{Connect to ACP and store a default client}
\usage{
acp_connect(url = "http://127.0.0.1:8765", token = NULL)
}
\arguments{
\item{url}{ACP base URL}

\item{token}{optional bearer token}
}
\value{
invisible TRUE
}
\description{
Connect to ACP and store a default client
}
