% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_api.R
\name{acp_get_default_client}
\alias{acp_get_default_client}
\title{Get the default ACP client if connected}
\usage{
acp_get_default_client()
}
\value{
ACP client object or NULL
}
\description{
Get the default ACP client if connected
}
