% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{acp_is_connected}
\alias{acp_is_connected}
\title{Determine whether an ACP client object appears valid}
\usage{
acp_is_connected(client)
}
\arguments{
\item{client}{object to inspect}
}
\value{
logical scalar
}
\description{
Determine whether an ACP client object appears valid
}
