% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.R
\name{acp_phenotype_recommendation_advice}
\alias{acp_phenotype_recommendation_advice}
\title{Call phenotype recommendation advice flow}
\usage{
acp_phenotype_recommendation_advice(client, study_intent)
}
\arguments{
\item{client}{ACP client object}

\item{study_intent}{non-empty study intent string}
}
\value{
parsed ACP response
}
\description{
Call phenotype recommendation advice flow
}
