% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{acp_server_info}
\alias{acp_server_info}
\title{Retrieve ACP service information}
\usage{
acp_server_info(client)
}
\arguments{
\item{client}{ACP client object}
}
\value{
Parsed \verb{/health} payload.
}
\description{
Retrieve ACP service information
}
