% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.R
\name{acp_suggest_cohort_method_specs}
\alias{acp_suggest_cohort_method_specs}
\title{Call cohort methods specification recommendation flow}
\usage{
acp_suggest_cohort_method_specs(
  client,
  study_intent,
  analytic_settings_description
)
}
\arguments{
\item{client}{ACP client object}

\item{study_intent}{protocol context string}

\item{analytic_settings_description}{free-text analytic settings description}
}
\value{
parsed ACP response
}
\description{
Call cohort methods specification recommendation flow
}
