% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.R
\name{acp_workflow_context_dialogue}
\alias{acp_workflow_context_dialogue}
\title{Call workflow context dialogue flow}
\usage{
acp_workflow_context_dialogue(client, stage_context, message)
}
\arguments{
\item{client}{ACP client object}

\item{stage_context}{workflow-stage context object}

\item{message}{latest user message}
}
\value{
parsed ACP response
}
\description{
Call workflow context dialogue flow
}
