% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint_and_concept_sets.R
\name{previewConceptSetPatch}
\alias{previewConceptSetPatch}
\title{Preview concept set patch}
\usage{
previewConceptSetPatch(conceptSetRef, patch)
}
\arguments{
\item{conceptSetRef}{path or URL}

\item{patch}{patch object from proposeIncludeDescendantsPatch}
}
\value{
preview result
}
\description{
Preview concept set patch
}
