% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint_and_concept_sets.R
\name{proposeIncludeDescendantsPatch}
\alias{proposeIncludeDescendantsPatch}
\title{Propose includeDescendants patch for concept set}
\usage{
proposeIncludeDescendantsPatch(conceptSetRef)
}
\arguments{
\item{conceptSetRef}{path or URL to concept_set.json}
}
\value{
list patch payload
}
\description{
Propose includeDescendants patch for concept set
}
