% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_api.R
\name{pullPhenotypeDefinitions}
\alias{pullPhenotypeDefinitions}
\title{Pull phenotype definitions to a local folder}
\usage{
pullPhenotypeDefinitions(cohortIds, outputDir = ".", overwrite = FALSE)
}
\arguments{
\item{cohortIds}{character vector of ACP phenotype ids}

\item{outputDir}{directory to write JSON definitions}

\item{overwrite}{logical; if FALSE, auto-version the filename}
}
\value{
character vector of written file paths
}
\description{
Pull phenotype definitions to a local folder
}
