% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_api.R
\name{selectPhenotypeRecommendations}
\alias{selectPhenotypeRecommendations}
\title{Select phenotype recommendations}
\usage{
selectPhenotypeRecommendations(
  recommendations,
  select = NULL,
  interactive = interactive()
)
}
\arguments{
\item{recommendations}{list from suggestPhenotypes()$phenotype_recommendations}

\item{select}{phenotype ids, integer indices, or "all"/NULL}

\item{interactive}{if TRUE and select is NULL, prompt user}
}
\value{
character vector of chosen phenotype ids
}
\description{
Select phenotype recommendations
}
